requirements.txt
django
django-ckeditor
pillow
gunicorn
django-heroku